# Add bounds checks to loops