@lukechampine 2020-01-24
Interesting. BCE seems to work in some cases here, but not others.

These asserts eliminate the bounds checking inside the loop:

https://github.com/lukechampine/fastxor/blob/43ff72405cae657665a5a017aaf36d0782f2a186/xor_amd64.go#L56-L61

but these do not:

https://github.com/lukechampine/fastxor/blob/43ff72405cae657665a5a017aaf36d0782f2a186/xor_amd64.go#L64-L69

The same is true of `Bytes`. This is still an improvement, though, because the former loops run much more frequently than the latter.

Strangely, I can't seem to eliminate the bounds checking in the non-amd64 implementations. In the original code:

https://github.com/lukechampine/fastxor/blob/7d8727055c51c4edee96a1a827a682a2a77f69b4/xor.go#L15-L17

there are bounds checks on `dst[i]`, `a[i]`, and `b[i]`, as expected. But in the PR code:

https://github.com/lukechampine/fastxor/blob/43ff72405cae657665a5a017aaf36d0782f2a186/xor.go#L15-L20

there's still one bounds check, on `a[i]`. Same deal in `Byte`; we can eliminate one of the bounds checks in the loop, but not both. Perhaps the compiler isn't smart enough to deal with loops that access multiple slices?

In any case, this a clear improvement over the status quo, so I'm happy to merge it. Perhaps as a follow-up, you could experiment with an explicit `wasm` build tag. If `wasm` could be implemented with the `xor_unaligned.go` code, you'd probably see a big speedup.