@lukechampine 2020-01-24
btw, I'm curious -- what are you using this code for? :)