@rubenv 2020-01-24
Cool, thanks!

It's part of the start-up I'm preparing. I'm using thi module for XOR-compressing repeated telemetry samples coming out of race cars. High frequency sampling where most numbers change only gradually and thus turn into a long string of zeroes: add a bit of run-length-encoding and it compresses like magic.

The annoying thing with wasm (and go) is that it's hard to benchmark. I'm not entirely convinced yet. Nonetheless: the server (and desktop) side of things will definitely be in Go, so all of this is definitely useful.

Thanks *a lot* for building it.